+++ aquadoggo/src/lib.rs
@@ -18,16 +18,17 @@ mod config;
 mod context;
 mod db;
 mod errors;
-/// Aquadoggo graphql types for handling client and replication requests
 pub mod graphql;
 mod http;
 mod manager;
 mod materializer;
 mod node;
+mod replication;
 
 #[cfg(test)]
 mod test_helpers;
 
-pub use config::Configuration;
+pub use crate::config::Configuration;
+pub use crate::replication::ReplicationConfiguration;
 pub use db::{connection_pool, provider::SqlStorage};
 pub use node::Node;